+++ CHANGELOG.md
@@ -330,3 +330,4 @@
 - [2026-04-12 09:44] test(network): add unit tests for network business logic
 - [2026-04-12 10:05] chore(theme): update dependencies and cleanup unused imports in theme
 - [2026-04-12 10:26] feat(storage): add async storage stream provider
+- [2026-04-12 11:12] test(navigation): implement widget integration tests for navigation